Problem Description
How you’ve approached implementing the ticket?
After reading through the description and the requirements of the client I started by first implementing the base HTML code and then adding styling and comments after all the content that is require has been added. If there were no design concepts or the client did not provide with a design brief/example I would have created a mock design on my notepad.

Any feedback/questions you had about ticket requirements?
There were a couple of grammar errors on the designs that were made on the Invision Link. And as far as functionality I did not get very far. Main questions I would have had was
"Assuming the database of email addresses, where those were held or are they given?"

What changes you think should be made to improve the UX?
As far as user experience goes it seems to do well as there is not much of a journey for the user to experience. The background colour to the form helps focus the user's attention and also allows those with sight problems to use with ease. The reset button is obvious and is clearly named, same with the "go back" button as it does not defer the user's attention but is large enough for the user to notice and easy enough to click. Overall UX is simple by design and usable.

What improvements you’d make to the component given more time, and give an estimate for how long those changes would take to implement?
I would add more functionality and also add animations a small animation for the help text where there is a countersink symbol. A simple aesthetic animation that tells the user that information is displayed below and can be hidden again where the countersink symbol faces up. A small transition animation for hovers on the buttons. Overall I believe adding the functionality and the small aesthetic animations could take at minimum 2 hours.